refactor(chrome): extract desktop chrome from app.rs
Move sidebar, timeline bar, chips/toasts/banners, the action registry and
the two remaining inline windows out of app.rs into app/chrome/, mirroring
app/mobile/. Bodies are unchanged; only visibility and imports differ.
