+++ crates/hookecho/src/app.rs
@@ -6,6 +6,7 @@
 /// Touch-first Android chrome (top bar, bottom dock, slide-up sheets), replacing the desktop
 /// drawer / pills / alert dock. Only the chrome differs; the map,
 /// windows, and every data path are shared.
+mod chrome;
 mod mobile;
 
 use crate::colormap::{ColorTable, Palettes};
@@ -3801,22 +3802,6 @@ impl HookEchoApp {
         });
     }
 
-    /// Show a warning banner, or refresh the matching one already on screen — a repeated event
-    /// bumps its clock instead of stacking a duplicate card over the radar.
-    fn banner(&mut self, event: String, area: String) {
-        // ponytail: linear scan; the lane holds a handful of cards at most.
-        if let Some(b) = self
-            .warning_banners
-            .iter_mut()
-            .find(|(e, _, _)| *e == event)
-        {
-            b.1 = area;
-            b.2 = Instant::now();
-        } else {
-            self.warning_banners.push((event, area, Instant::now()));
-        }
-    }
-
     /// Height of pane `idx`'s beam centre above the radar, in feet, over the point `ll`
     /// (`[lon, lat]`). `None` when the pane has no site or no loaded tilt.
     ///
@@ -3990,124 +3975,6 @@ impl HookEchoApp {
                 self.toast(ToastKind::Error, format!("{feed} unavailable — {err}"));
             }
         }
-    }
-
-    /// Say something happened. Operation results used to reach the user only through the log.
-    fn toast(&mut self, kind: ToastKind, text: impl Into<String>) {
-        self.toasts.push(Toast {
-            text: text.into(),
-            kind,
-            at: Instant::now(),
-        });
-    }
-
-    /// Toast stack, below the right-hand control column. Expires after ~4 s, fading out over the
-    /// last second; click to dismiss.
-    /// ponytail: clock-based fade, no per-toast animation ids.
-    fn show_toasts(&mut self, ctx: &egui::Context) {
-        const LIFE: f32 = 4.0;
-        self.toasts.retain(|t| t.at.elapsed().as_secs_f32() < LIFE);
-        if self.toasts.is_empty() {
-            return;
-        }
-        ctx.request_repaint_after(std::time::Duration::from_millis(100));
-        let accent = crate::theme::accent(self.settings.theme);
-        let mut dismiss = None;
-        egui::Area::new(egui::Id::new("toasts"))
-            .constrain_to(self.chrome_rect)
-            .anchor(
-                egui::Align2::RIGHT_TOP,
-                egui::vec2(
-                    crate::ui::style::LANE_RIGHT_BADGE_X,
-                    crate::ui::style::lane_right_badge_y(6),
-                ),
-            )
-            .show(ctx, |ui| {
-                for (i, t) in self.toasts.iter().enumerate() {
-                    let left = LIFE - t.at.elapsed().as_secs_f32();
-                    let alpha = left.min(1.0).clamp(0.0, 1.0);
-                    let stripe = match t.kind {
-                        ToastKind::Info => accent,
-                        ToastKind::Success => crate::ui::style::OMEGA_GREEN,
-                        ToastKind::Error => egui::Color32::from_rgb(220, 90, 90),
-                    };
-                    let resp = crate::ui::style::glass(ui, (220.0 * alpha) as u8)
-                        .stroke(egui::Stroke::new(1.0, stripe.gamma_multiply(alpha)))
-                        .show(ui, |ui| {
-                            ui.set_max_width(320.0);
-                            ui.label(
-                                egui::RichText::new(&t.text)
-                                    .size(crate::ui::style::FONT_BASE)
-                                    .color(egui::Color32::from_gray(235).gamma_multiply(alpha)),
-                            );
-                        })
-                        .response;
-                    if resp.interact(egui::Sense::click()).clicked() {
-                        dismiss = Some(i);
-                    }
-                    ui.add_space(6.0);
-                }
-            });
-        if let Some(i) = dismiss {
-            self.toasts.remove(i);
-        }
-    }
-
-    /// Draw new-warning banners at top-center (auto-expire ~45s; click to dismiss all).
-    fn show_warning_banners(&mut self, ctx: &egui::Context) {
-        self.warning_banners
-            .retain(|(_, _, at)| at.elapsed().as_secs() < 45);
-        if self.warning_banners.is_empty() {
-            return;
-        }
-        egui::Area::new(egui::Id::new("warning_banners"))
-            .constrain_to(self.chrome_rect)
-            // Read-only banner that self-expires in 45 s. Interactable layers occlude the map's
-            // pinch test (`layer_id_at`), and a banner across the top of a phone screen is exactly
-            // where a two-finger gesture lands, so it must not take input.
-            .interactable(false)
-            .anchor(
-                egui::Align2::CENTER_TOP,
-                egui::vec2(0.0, crate::ui::style::LANE_TOP_BANNER),
-            )
-            .show(ctx, |ui| {
-                for (event, area, at) in &self.warning_banners {
-                    // Fade out over the last two seconds instead of vanishing mid-read.
-                    let a = ((45.0 - at.elapsed().as_secs_f32()) / 2.0).clamp(0.0, 1.0);
-                    egui::Frame::new()
-                        .fill(egui::Color32::from_rgb(150, 20, 20).gamma_multiply(a))
-                        .stroke(egui::Stroke::new(
-                            1.0,
-                            egui::Color32::from_rgb(255, 120, 120).gamma_multiply(a),
-                        ))
-                        .corner_radius(egui::CornerRadius::same(6))
-                        .inner_margin(egui::Margin::symmetric(12, 6))
-                        .show(ui, |ui| {
-                            ui.horizontal(|ui| {
-                                ui.label(
-                                    egui::RichText::new(egui_phosphor::regular::WARNING)
-                                        .size(16.0)
-                                        .color(egui::Color32::WHITE.gamma_multiply(a)),
-                                );
-                                ui.vertical(|ui| {
-                                    ui.label(
-                                        egui::RichText::new(format!("New {event}"))
-                                            .strong()
-                                            .color(egui::Color32::WHITE.gamma_multiply(a)),
-                                    );
-                                    if !area.is_empty() {
-                                        ui.label(egui::RichText::new(area).small().color(
-                                            egui::Color32::from_gray(230).gamma_multiply(a),
-                                        ));
-                                    }
-                                });
-                            });
-                        });
-                    ui.add_space(4.0);
-                }
-            });
-        // Fast enough for the fade-out; the banner is only up for 45 s.
-        ctx.request_repaint_after(std::time::Duration::from_millis(100));
     }
 
     /// Detect warning-tier alerts whose id we haven't seen, raising a banner + audible cue for
@@ -6161,121 +6028,6 @@ impl HookEchoApp {
             .count()
     }
 
-    /// The tornado-climatology results window: a magnitude histogram + strongest-first list of
-    /// historical tornadoes near the clicked point.
-    fn show_climatology_window(&mut self, ctx: &egui::Context) {
-        if !self.climo_open {
-            return;
-        }
-        let mut open = self.climo_open;
-        crate::ui::phone_surface(ctx, egui::Window::new("Tornado climatology"))
-            .open(&mut open)
-            .default_width(360.0)
-            .show(ctx, |ui| {
-                if let Some((lon, lat)) = self.climo_center {
-                    ui.label(format!("Within 25 mi of {lat:.3}, {lon:.3}"));
-                }
-                if self.climo_loading {
-                    ui.horizontal(|ui| {
-                        ui.spinner();
-                        ui.label("Loading SPC tornado database (1950–2022)…");
-                    });
-                } else if let Some(e) = &self.climo_error {
-                    ui.colored_label(
-                        egui::Color32::from_rgb(230, 90, 90),
-                        format!("Load failed: {e}"),
-                    );
-                } else {
-                    ui.horizontal(|ui| {
-                        ui.strong(format!("{} tornadoes on record", self.climo_hits.len()));
-                        ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-                            let hits = &self.climo_hits;
-                            crate::ui::csv_buttons(
-                                ui,
-                                "tornadoes.csv",
-                                "Every tornado on record here, not just the first 50",
-                                || {
-                                    let mut s = String::from(
-                                        "year,mag,start_lat,start_lon,end_lat,end_lon\n",
-                                    );
-                                    for t in hits {
-                                        s.push_str(&format!(
-                                            "{},{},{:.4},{:.4},{:.4},{:.4}\n",
-                                            t.year, t.mag, t.slat, t.slon, t.elat, t.elon
-                                        ));
-                                    }
-                                    s
-                                },
-                            );
-                        });
-                    });
-                    let hist = wxdata::torclimo::mag_histogram(&self.climo_hits);
-                    ui.horizontal_wrapped(|ui| {
-                        for (i, label) in ["EF0", "EF1", "EF2", "EF3", "EF4", "EF5", "Unk"]
-                            .iter()
-                            .enumerate()
-                        {
-                            crate::theme::stat_card(ui, label, &hist[i].to_string());
-                        }
-                    });
-                    ui.separator();
-                    egui::ScrollArea::vertical()
-                        .max_height(240.0)
-                        .show(ui, |ui| {
-                            for t in self.climo_hits.iter().take(50) {
-                                let mag = if t.mag < 0 {
-                                    "EF?".to_string()
-                                } else {
-                                    format!("EF{}", t.mag)
-                                };
-                                ui.label(format!(
-                                    "{}  {}  start {:.2},{:.2}",
-                                    t.year, mag, t.slat, t.slon
-                                ));
-                            }
-                            if self.climo_hits.len() > 50 {
-                                ui.weak(format!("… and {} more", self.climo_hits.len() - 50));
-                            }
-                        });
-                }
-                ui.separator();
-                ui.strong("Warning history");
-                ui.weak("How often this spot's county has been warned (IEM, 1986–present).");
-                match (&self.climo_warn, self.climo_warn_rx.is_some()) {
-                    (Some(s), _) if s.total == 0 => {
-                        ui.label("No warnings on record here.");
-                    }
-                    (Some(s), _) => {
-                        ui.horizontal_wrapped(|ui| {
-                            crate::theme::stat_card(ui, "Warnings", &s.total.to_string());
-                            if let Some(y) = s.first_year {
-                                crate::theme::stat_card(ui, "Since", &y.to_string());
-                            }
-                            if let Some((y, n)) = s.busiest_year {
-                                crate::theme::stat_card(ui, "Busiest year", &format!("{y} ({n})"));
-                            }
-                            if let Some((d, n)) = s.worst_day {
-                                crate::theme::stat_card(ui, "Worst day", &format!("{d} ({n})"));
-                            }
-                        });
-                        for (name, n) in s.by_name.iter().take(8) {
-                            ui.label(format!("{n} × {name}"));
-                        }
-                    }
-                    (None, true) => {
-                        ui.horizontal(|ui| {
-                            ui.spinner();
-                            ui.label("Loading warning history…");
-                        });
-                    }
-                    (None, false) => {
-                        ui.weak("Warning history unavailable.");
-                    }
-                }
-            });
-        self.climo_open = open;
-    }
-
     /// Tornado-climatology query at `(lon, lat)`: if the SPC track database is loaded, list nearby
     /// historical tornadoes; otherwise start the (cached) async load and queue the query.
     fn query_climatology(&mut self, lon: f64, lat: f64) {
@@ -6774,222 +6526,6 @@ impl HookEchoApp {
         (rows.len(), max_esc)
     }
 
-    /// The sidebar: everything that isn't the map.
-    ///
-    /// A docked left column holding the whole action registry (products, layers, tools,
-    /// windows — searchable) with the app's own commands below it. It replaces a wall of parallel
-    /// entry points; the map keeps nothing but transient status.
-    fn sidebar(&mut self, root: &mut egui::Ui, ctx: &egui::Context) {
-        crate::prof_scope!("sidebar");
-        let accent = crate::theme::accent(self.settings.theme);
-        let entries = self.palette_entries();
-        let mut query = std::mem::take(&mut self.layers_query);
-        let (mut chosen, mut fly_to) = (None, None);
-        let mut opts = ui::layer_options::UiActions::default();
-        let mut focus_search = std::mem::take(&mut self.sidebar_focus_search);
-        let mut alerts_tab = self.show_alert_panel;
-        let (alert_count, _) = self.alert_badge();
-        let bounds = self.view_bounds();
-        let feats = self.active_alert_features().to_vec();
-        let mut muted = self.settings.mute_alerts;
-        let mut alert_hit = None;
-        // Read before the panel closure: the Layer options callback runs inside a `&mut self`
-        // borrow and can only touch plain fields, not `&self` methods.
-        let l3_site = self.l3grid_site.clone();
-        let tz = self.active_tz();
-        let mosaic = self.mosaic_status();
-        let mut etop_dbz = self.settings.etop_dbz;
-        let mut hide = false;
-        // Tour spotlights, recorded from inside the panel (no `self` reachable in there).
-        let mut alerts_rect = None;
-        let sidebar_rect = egui::Panel::left("sidebar")
-            .exact_size(264.0)
-            .show(root, |ui| {
-                // Title on the same line as the tabs: the name is branding, not a section, and
-                // its own row plus separator cost 30 px of every screen height.
-                ui.horizontal(|ui| {
-                    ui.label(
-                        egui::RichText::new("Hook Echo-WX")
-                            .size(13.0)
-                            .strong()
-                            .color(accent),
-                    );
-                    ui.add_space(4.0);
-                    // Data | Alerts. `show_alert_panel` is the same flag the bell and the A hotkey
-                    // flip, so every entry point lands on the same tab.
-                    if ui.selectable_label(!alerts_tab, "Data").clicked() {
-                        alerts_tab = false;
-                    }
-                    let label = if alert_count == 0 {
-                        "Alerts".to_string()
-                    } else {
-                        format!("Alerts ({alert_count})")
-                    };
-                    let alerts_hit = ui.selectable_label(alerts_tab, label);
-                    alerts_rect = Some(alerts_hit.rect);
-                    if alerts_hit.clicked() {
-                        alerts_tab = true;
-                    }
-                    // Collapse, on the row it collapses. The floating button that brings the
-                    // panel back lands in the same corner this one sits in.
-                    ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-                        if ui
-                            .add(
-                                egui::Button::new(
-                                    egui::RichText::new(egui_phosphor::regular::CARET_LEFT)
-                                        .size(14.0),
-                                )
-                                .fill(egui::Color32::TRANSPARENT)
-                                .stroke(egui::Stroke::NONE),
-                            )
-                            .on_hover_text("Hide the sidebar")
-                            .clicked()
-                        {
-                            hide = true;
-                        }
-                    });
-                });
-                ui.separator();
-                if alerts_tab {
-                    alert_hit = ui::alert_panel::body(ui, &feats, bounds, &mut muted);
-                    return;
-                }
-                self.product_section(ui, &mut opts);
-                ui.separator();
-                // A drag rewrites the order in place, so persist it when it moves.
-                let order_was = self.settings.layer_order.clone();
-                chosen = ui::layers_panel::body(
-                    ui,
-                    &entries,
-                    &mut query,
-                    accent,
-                    // Leave room for the disclosures under the tree, whatever the window height.
-                    (ui.available_height() - 110.0).max(120.0),
-                    std::mem::take(&mut focus_search),
-                    &mut self.settings.layer_order,
-                    |ui| {
-                        // Knobs for the layers that are already on, drawn between the Radar group
-                        // and the rest. Collapsed by default: the list is still the panel's job.
-                        egui::CollapsingHeader::new("Layer options")
-                            .default_open(false)
-                            .show(ui, |ui| {
-                                let glm_options = self.show_glm
-                                    || self.fields.get(&crate::render::FieldLayer::GlmFed).is_some_and(|s| s.show);
-                                crate::ui::layer_options::show(
-                                    ui,
-                                    &mut self.filters,
-                                    &mut self.fields,
-                                    &mut self.rotation_minutes,
-                                    &mut self.hrrr_fcst_hour,
-                                    self.hrrr_valid,
-                                    tz,
-                                    &mut self.env_cape_ml,
-                                    &mut self.env_srh_km,
-                                    &mut self.env_model,
-                                    &mut self.contour_kind,
-                                    &mut etop_dbz,
-                                    &mut self.snow_hours,
-                                    &self.show_tropical,
-                                    &mut self.tropical_wind_kt,
-                                    &mut self.tropical_surge,
-                                    l3_site.as_deref(),
-                                    &mut self.global_model,
-                                    &mut self.global_fcst_hour,
-                                    &mut self.diff_field,
-                                    self.diff_valid.as_ref(),
-                                    &mut self.settings.lightning_minutes,
-                                    glm_options,
-                                    &mut self.settings.glm_goes_west,
-                                    self.show_spotters,
-                                    &mut self.settings.spotter_range_km,
-                                    &mut self.settings.detectors,
-                                    Some(mosaic.as_str()),
-                                    &mut opts,
-                                );
-                            });
-                    },
-                );
-                self.settings.etop_dbz = etop_dbz;
-                if self.settings.layer_order != order_was {
-                    self.settings.save();
-                }
-                // Place search folds in here rather than keeping a pill of its own: action
-                // matches rank first, and this row is the explicit "I meant a place" answer.
-                if !query.trim().is_empty() {
-                    ui.add_space(4.0);
-                    let w = ui.available_width();
-                    if ui
-                        .add(
-                            egui::Button::new(
-                                egui::RichText::new(format!(
-                                    "{}  Fly to \u{201c}{}\u{201d}",
-                                    egui_phosphor::regular::MAP_PIN,
-                                    query.trim()
-                                ))
-                                .size(13.0),
-                            )
-                            .min_size(egui::vec2(w, 34.0))
-                            .corner_radius(10.0),
-                        )
-                        .on_hover_text("Search the place name and move the map there")
-                        .clicked()
-                    {
-                        fly_to = Some(query.trim().to_string());
-                    }
-                }
-                ui.add_space(4.0);
-                // The set-once map knobs, and the app's own commands.
-                egui::CollapsingHeader::new("Map")
-                    .default_open(false)
-                    .show(ui, |ui| self.map_rows(ui, &mut opts));
-                egui::CollapsingHeader::new("App")
-                    .default_open(false)
-                    .show(ui, |ui| self.app_rows(ui));
-            })
-            .response
-            .rect;
-        self.tour_anchors.menu = Some(sidebar_rect);
-        self.tour_anchors.alerts = alerts_rect;
-        self.show_alert_panel = alerts_tab;
-        self.settings.mute_alerts = muted;
-        if hide {
-            self.settings.hide_sidebar = true;
-            self.settings.save();
-        }
-        if let Some((id, lon, lat)) = alert_hit {
-            // Fly the active camera to the alert and open its bulletin.
-            let cam = &mut self.views[self.active].camera;
-            cam.center = crate::render::mercator::lonlat_to_world(lon, lat);
-            cam.zoom = cam.zoom.max(8.0);
-            self.open_alert_popup(&id);
-        }
-        // `query` is the live text; `self.layers_query` was taken from at the top of the frame.
-        let searched = !query.trim().is_empty();
-        self.layers_query = query;
-        self.apply_ui_actions(opts, ctx);
-        if let Some(a) = chosen {
-            // Picking a search hit means you're done searching: clear the query so the tree comes
-            // back. Browsing the list without a query is the opposite — you're flipping layers on
-            // and off, so it stays put.
-            if searched || matches!(a, PaletteAction::OpenWindow(_)) {
-                self.layers_query.clear();
-            }
-            self.apply_palette(a, ctx);
-        }
-        if let Some(place) = fly_to {
-            self.geocode_nav = true;
-            self.save_offer = None; // a new search retires the previous offer
-            self.place_status = Some(("Searching…".to_string(), Instant::now()));
-            let http = self.http.clone();
-            let tx = self.geocode_tx.clone();
-            let ctx2 = ctx.clone();
-            self.spawner.spawn(async move {
-                let _ = tx.send(wxdata::geocode::search(&http, &place).await);
-                ctx2.request_repaint();
-            });
-        }
-    }
-
     /// Thin right-edge colorbar: the active pane's moment scale, docked so it never covers the map.
     /// Docked timeline scrubber (desktop): transport + scrub + live badge in a full-width bar
     /// under the map. The date picker, loop and speed are one right-click away on the
@@ -6998,270 +6534,6 @@ impl HookEchoApp {
     pub(crate) fn active_tz(&self) -> Option<wxdata::tz::Tz> {
         self.settings
             .tz_for(self.views[self.active].site.as_deref())
-    }
-
-    fn timeline_bar(&mut self, root: &mut egui::Ui) {
-        crate::prof_scope!("timeline_bar");
-        use egui_phosphor::regular as ph;
-        let accent = crate::theme::accent(self.settings.theme);
-        let tz = self.active_tz();
-        let fresh = self.views[self.active]
-            .volume
-            .as_ref()
-            .is_some_and(|v| (chrono::Utc::now() - v.time).num_seconds() < 900);
-        // Site and data age used to live in the docked status bar; the clock belongs with the clock.
-        let site = self.views[self.active]
-            .site
-            .clone()
-            .unwrap_or_else(|| "no site".to_string());
-        let age = self.views[self.active].volume.as_ref().map(|v| {
-            let secs = (Utc::now() - v.time).num_seconds().max(0);
-            format!("({} ago)", humanize(secs))
-        });
-        let loading = self.views[self.active].loading;
-        let mut go_head = false;
-        // Soonest rain arrival and the DVR buffer depth ride the pill: both are about time, and
-        // both used to sit in an always-on chip in the opposite corner.
-        let rain = self
-            .rain_eta
-            .iter()
-            .min_by(|a, b| a.1.partial_cmp(&b.1).unwrap_or(std::cmp::Ordering::Equal))
-            .map(|(name, min)| format!("\u{1f327} {name} ~{min:.0} min"));
-        let dvr = self.dvr_depth();
-        // Edited through a local so the pill closure keeps its single `&mut self.views` borrow.
-        let mut loop_frames = self.settings.live_loop_frames;
-        // Where the scrubber lands, for the tour's spotlight (same reason: no `self` in there).
-        let mut scrub_rect = None;
-        egui::Panel::bottom("timeline_bar")
-            .exact_size(44.0)
-            .show(root, |ui| {
-                let t = &mut self.views[self.active].timeline;
-                ui.horizontal_centered(|ui| {
-                    ui.label(
-                        egui::RichText::new(&site)
-                            .size(crate::ui::style::FONT_BASE)
-                            .strong()
-                            .color(egui::Color32::from_gray(238)),
-                    );
-                    let btn = |ui: &mut egui::Ui, glyph: &str, on: bool| {
-                        let fg = if on {
-                            accent
-                        } else {
-                            egui::Color32::from_gray(225)
-                        };
-                        ui.add(
-                            egui::Button::new(egui::RichText::new(glyph).size(18.0).color(fg))
-                                .min_size(egui::vec2(30.0, 30.0))
-                                .fill(egui::Color32::TRANSPARENT)
-                                .stroke(egui::Stroke::NONE),
-                        )
-                        .clicked()
-                    };
-                    if btn(ui, ph::SKIP_BACK, false) {
-                        t.step(-1);
-                    }
-                    let playing = t.playing;
-                    if btn(ui, if playing { ph::PAUSE } else { ph::PLAY }, playing) {
-                        t.toggle_play();
-                    }
-                    if btn(ui, ph::SKIP_FORWARD, false) {
-                        t.step(1);
-                    }
-                    // Live / archive badge: click to re-pin to the newest volume.
-                    //
-                    // Pinned to live but the newest volume is old means the site's feed has
-                    // stopped, not that playback drifted — there is nothing newer to jump to.
-                    // The colour said so already; saying "LIVE" over a volume hours old did not.
-                    let (col, text, hint) = if t.following && fresh {
-                        (
-                            mobile::OMEGA_GREEN,
-                            "LIVE".to_string(),
-                            "Following the newest volume.",
-                        )
-                    } else if t.following {
-                        (
-                            egui::Color32::from_rgb(220, 180, 0),
-                            "STALE".to_string(),
-                            "Following the newest volume, but this site has not produced one \
-                             recently — its feed has stopped. The age next to the clock is how \
-                             far behind it is.",
-                        )
-                    } else {
-                        (
-                            egui::Color32::from_gray(150),
-                            format!("ARCHIVE {}", t.date.format("%m/%d")),
-                            "Scrubbed to an archive day. Click to jump back to live.",
-                        )
-                    };
-                    let badge = ui.add(
-                        egui::Button::new(
-                            egui::RichText::new(text)
-                                .size(12.0)
-                                .strong()
-                                .color(egui::Color32::BLACK),
-                        )
-                        .fill(col)
-                        .corner_radius(9.0),
-                    )
-                    .on_hover_text(hint);
-                    if badge.clicked() {
-                        go_head = true;
-                    }
-                    // Right-click the badge for the knobs that used to sit in the toolbox's
-                    // Timeline section: which archive day, and how playback loops.
-                    egui::Popup::context_menu(&badge)
-                        .align(egui::RectAlign::TOP_START)
-                        .show(|ui| {
-                            ui.set_min_width(240.0);
-                            ui.horizontal(|ui| {
-                                ui.label("Date:");
-                                if ui.button(egui_phosphor::regular::CARET_LEFT).clicked() {
-                                    if let Some(d) = t.date.pred_opt() {
-                                        t.date = d;
-                                        t.following = false;
-                                    }
-                                }
-                                // The carets are still the fastest way to step a day, but they
-                                // were the *only* way: reaching a storm from years back meant
-                                // thousands of clicks. The archive runs to June 1991.
-                                let today = chrono::Utc::now().date_naive();
-                                if let Some(d) = archive_day_input(ui, t.date) {
-                                    // Clamp rather than trust the input: neither a calendar
-                                    // widget nor a typed string knows where the archive starts
-                                    // or that the future is empty.
-                                    t.date = d.clamp(wxdata::level2::ARCHIVE_START, today);
-                                    t.following = t.date >= today;
-                                }
-                                let is_today = t.date >= chrono::Utc::now().date_naive();
-                                if ui
-                                    .add_enabled(
-                                        !is_today,
-                                        egui::Button::new(egui_phosphor::regular::CARET_RIGHT),
-                                    )
-                                    .clicked()
-                                {
-                                    if let Some(d) = t.date.succ_opt() {
-                                        t.date = d;
-                                    }
-                                }
-                            });
-                            ui.horizontal(|ui| {
-                                if ui.button("⏮").on_hover_text("First frame").clicked() {
-                                    t.go_begin();
-                                }
-                                ui.checkbox(&mut t.loop_enabled, "Loop");
-                            });
-                            ui.add(
-                                egui::Slider::new(&mut t.speed, 1.0..=15.0)
-                                    .suffix(" fps")
-                                    .show_value(true),
-                            );
-                            ui.add(
-                                egui::DragValue::new(&mut loop_frames)
-                                    .range(2..=30)
-                                    .suffix(" frames"),
-                            )
-                            .on_hover_text(
-                                "How many of the newest volumes ▶ cycles through when live",
-                            );
-                        });
-                    // Scrub bar + readout fill the rest of the pill.
-                    let observed = t.frames.len();
-                    if observed == 0 {
-                        ui.weak(if t.listing {
-                            "listing volumes…"
-                        } else {
-                            "(no volumes)"
-                        });
-                        return;
-                    }
-                    let readout = match t.forecast_hour() {
-                        Some(h) => format!("F+{h}h"),
-                        None => t
-                            .current()
-                            .and_then(|id| id.date_time())
-                            .map(|d| crate::timefmt::fmt_clock(d, tz, true))
-                            .unwrap_or_default(),
-                    };
-                    let last = t.slot_count().saturating_sub(1);
-                    let mut ph_idx = t.playhead;
-                    let slider_w = (ui.available_width() - 92.0).max(80.0);
-                    let resp = ui.add_sized(
-                        [slider_w, 20.0],
-                        egui::Slider::new(&mut ph_idx, 0..=last).show_value(false),
-                    );
-                    scrub_rect = Some(resp.rect);
-                    if resp.changed() {
-                        t.playhead = ph_idx;
-                        t.playing = false;
-                        t.following = ph_idx + 1 == observed;
-                    }
-                    // Mark where observed radar ends and the HRRR forecast tail begins, so a
-                    // scrub past the head reads as a model run and not as more radar.
-                    if t.slot_count() > observed {
-                        let frac = observed as f32 / t.slot_count() as f32;
-                        let r = resp.rect;
-                        let x = r.left() + frac * r.width();
-                        let p = ui.painter_at(r);
-                        p.rect_filled(
-                            egui::Rect::from_min_max(
-                                egui::pos2(x, r.top()),
-                                egui::pos2(r.right(), r.bottom()),
-                            ),
-                            0.0,
-                            // Faint enough to stay behind the slider, solid enough to read on
-                            // the dark bar — at alpha 36 it was invisible in a screenshot.
-                            egui::Color32::from_rgba_unmultiplied(120, 170, 240, 90),
-                        );
-                        p.line_segment(
-                            [egui::pos2(x, r.top()), egui::pos2(x, r.bottom())],
-                            egui::Stroke::new(1.0, egui::Color32::from_rgb(120, 170, 240)),
-                        );
-                    }
-                    ui.label(
-                        egui::RichText::new(readout)
-                            .size(12.0)
-                            .monospace()
-                            .color(egui::Color32::from_gray(215)),
-                    );
-                    if let Some(r) = &rain {
-                        ui.label(
-                            egui::RichText::new(r)
-                                .size(crate::ui::style::FONT_SM)
-                                .color(egui::Color32::from_rgb(110, 180, 240)),
-                        )
-                        .on_hover_text(
-                            "Estimated from storm motion \u{2014} rough for backbuilding storms",
-                        );
-                    }
-                    if dvr > 1 {
-                        ui.label(
-                            egui::RichText::new(format!("\u{27f2} {dvr}"))
-                                .size(crate::ui::style::FONT_SM)
-                                .color(egui::Color32::from_gray(150)),
-                        )
-                        .on_hover_text("Frames buffered in memory for instant replay (R)");
-                    }
-                    if let Some(age) = &age {
-                        ui.label(
-                            egui::RichText::new(age)
-                                .size(crate::ui::style::FONT_SM)
-                                .color(egui::Color32::from_gray(150)),
-                        );
-                    } else if loading {
-                        ui.label(
-                            egui::RichText::new("loading…")
-                                .size(crate::ui::style::FONT_SM)
-                                .color(egui::Color32::from_gray(150)),
-                        );
-                    }
-                });
-            });
-        self.settings.live_loop_frames = loop_frames;
-        self.tour_anchors.timeline = scrub_rect;
-        if go_head {
-            self.views[self.active].timeline.go_head();
-        }
     }
 
     /// Sidebar header: the site, what you're looking at, its tilt, and the per-product knobs.
@@ -7620,959 +6892,6 @@ impl HookEchoApp {
             T::MiniLoop => &mut self.mini_loop,
             T::Blockage => &mut self.show_blockage,
         }
-    }
-
-    /// Every layer/product/tool/window as a searchable, categorized row. Consumed by the layers
-    /// panel (desktop slide-in + mobile sheet) and the Ctrl+K command palette.
-    /// The action registry, rebuilt at most once a frame.
-    ///
-    /// Building it allocates ~150 owned strings, and up to five places render from it in the same
-    /// frame (drawer, mobile sheets, legend, palette). Nothing can change it mid-frame — an
-    /// action dispatched from one of those lists takes effect on the next one — so a per-frame
-    /// memo is both free and honest.
-    /// The command-palette action registry for this frame. Shared, not copied: the built list is
-    /// a few hundred owned Strings, and several call sites want it in the same frame.
-    pub(crate) fn palette_entries(&mut self) -> std::sync::Arc<[PaletteEntry]> {
-        if let Some((frame, entries)) = &self.palette_cache {
-            if *frame == self.frame_nr {
-                return std::sync::Arc::clone(entries);
-            }
-        }
-        let entries: std::sync::Arc<[PaletteEntry]> = self.palette_entries_build().into();
-        self.palette_cache = Some((self.frame_nr, std::sync::Arc::clone(&entries)));
-        entries
-    }
-
-    fn palette_entries_build(&mut self) -> Vec<PaletteEntry> {
-        crate::prof_scope!("palette_entries");
-        use crate::render::FieldLayer as FL;
-        use AppWindow as W;
-        use OverlayToggle as T;
-        let mut out = Vec::new();
-        // Reverse index from the live bindings, so a rebind relabels every row that shows a chip.
-        let keys: Vec<(PaletteAction, String)> = crate::hotkeys::active(&self.settings)
-            .iter()
-            .filter_map(|b| match b.action {
-                crate::hotkeys::BindableAction::Palette(p) => {
-                    Some((p, crate::hotkeys::pretty(&b.shortcut)))
-                }
-                _ => None,
-            })
-            .collect();
-        let mut push = |label: &str, category, desc, common, action, on| {
-            out.push(PaletteEntry {
-                label: label.to_string(),
-                category,
-                action,
-                on,
-                desc,
-                common,
-                key: keys
-                    .iter()
-                    .find(|(a, _)| *a == action)
-                    .map(|(_, k)| k.clone()),
-            })
-        };
-
-        // --- Radar products (the active pane's moment). ---
-        let (cur_moment, cur_srv) = {
-            let v = &self.views[self.active];
-            (v.moment, v.srv)
-        };
-        // Every product, plus the storm-relative variant of velocity.
-        let rows = crate::products::PRODUCTS
-            .iter()
-            .map(|p| (p.moment, false, p.name, p.blurb))
-            .chain([(
-                Moment::Velocity,
-                true,
-                "Storm-Relative Velocity",
-                "Velocity with the storm's own motion subtracted out",
-            )]);
-        let have = self.available_moments();
-        for (m, srv, label, desc) in rows {
-            // A product this radar doesn't send is absent, not a row that paints nothing.
-            if !have[m.index()] {
-                continue;
-            }
-            let on = cur_moment == m && (m != Moment::Velocity || cur_srv == srv);
-            push(
-                label,
-                "Radar",
-                desc,
-                true,
-                PaletteAction::SetMoment(m, srv),
-                Some(on),
-            );
-        }
-
-        // --- National / model grids. ---
-        for (layer, category, label, desc, common) in [
-            (
-                FL::Mrms,
-                "National",
-                "MRMS Mosaic",
-                "Every radar in the country stitched into one picture",
-                true,
-            ),
-            (
-                FL::Mosaic,
-                "National",
-                "Radar Mosaic",
-                "Every nearby radar's own base reflectivity, stitched seamlessly",
-                true,
-            ),
-            (
-                FL::Rotation,
-                "National",
-                "Rotation tracks",
-                "Where rotation has passed over the last hour — the tornado-track map",
-                true,
-            ),
-            (
-                FL::Mesh,
-                "National",
-                "MESH hail",
-                "Estimated largest hail size each storm is producing",
-                true,
-            ),
-            (
-                FL::Lightning,
-                "National",
-                "Lightning density (CG)",
-                "Ground strikes only: NLDN cloud-to-ground density, averaged over a window you \
-                 pick. Pair it with satellite lightning (GLM) to see total vs cloud-to-ground.",
-                true,
-            ),
-            (
-                FL::AzShear,
-                "National",
-                "AzShear (0–2 km)",
-                "Low-level rotation strength, right now",
-                false,
-            ),
-            (
-                FL::PrecipRate,
-                "National",
-                "Rain rate",
-                "How hard it is coming down right now, rather than how much has fallen",
-                false,
-            ),
-            (
-                FL::Qpe1h,
-                "National",
-                "QPE 1-hour",
-                "How much rain has fallen in the last hour",
-                false,
-            ),
-            (
-                FL::Qpe24h,
-                "National",
-                "QPE 24-hour",
-                "How much rain has fallen in the last day",
-                false,
-            ),
-            (
-                FL::PrecipType,
-                "National",
-                "Precip type",
-                "Rain, snow, sleet or freezing rain at the surface",
-                false,
-            ),
-            (
-                FL::FlashFlood,
-                "National",
-                "FLASH flood ARI",
-                "How rare this much rain is here — flash-flood risk",
-                false,
-            ),
-            (
-                FL::HailSwath,
-                "National",
-                "Hail swaths (24 h)",
-                "Where hail has fallen over the past day",
-                false,
-            ),
-            (
-                FL::Vil,
-                "National",
-                "Digital VIL (L3)",
-                "How much water the storm is holding aloft",
-                false,
-            ),
-            (
-                FL::EchoTops,
-                "National",
-                "Echo tops (L3)",
-                "How tall the storm is",
-                false,
-            ),
-            (
-                FL::Hca,
-                "National",
-                "Hydrometeor class (L3)",
-                "What the radar thinks it's seeing: rain, hail, debris",
-                false,
-            ),
-            (
-                FL::GlmFed,
-                "National",
-                "Flash density (GLM)",
-                "Where the satellite flashes are densest \u{2014} the total-lightning field                  behind the individual dots, and where a lightning jump shows up first.",
-                false,
-            ),
-            (
-                FL::CompositeLocal,
-                "Radar",
-                "Composite reflectivity",
-                "The strongest echo anywhere above each point, not just what this tilt cuts through",
-                false,
-            ),
-            (
-                FL::VilLocal,
-                "Radar",
-                "VIL (derived)",
-                "Water held aloft, computed from this volume \u{2014} works in archive replay",
-                false,
-            ),
-            (
-                FL::VilDensity,
-                "Radar",
-                "VIL density (derived)",
-                "Water aloft per unit storm depth \u{2014} high values mean large hail",
-                false,
-            ),
-            (
-                FL::EtopLocal,
-                "Radar",
-                "Echo tops (derived)",
-                "Storm-top height at a threshold you pick, from this volume",
-                false,
-            ),
-            (
-                FL::HailMehs,
-                "Radar",
-                "Max hail size (derived)",
-                "Largest hail this storm can be making, from the volume aloft (live only)",
-                false,
-            ),
-            (
-                FL::HailPosh,
-                "Radar",
-                "Severe hail probability (derived)",
-                "Odds this storm is producing hail an inch or larger (live only)",
-                false,
-            ),
-            (
-                FL::GlobalMslp,
-                "Models",
-                "Global MSLP",
-                "Surface pressure worldwide — GFS or ECMWF, your pick in Layer options",
-                false,
-            ),
-            (
-                FL::GlobalHeight500,
-                "Models",
-                "Global 500 hPa height",
-                "The steering flow: where the troughs and ridges are, worldwide",
-                false,
-            ),
-            (
-                FL::GlobalTemp2m,
-                "Models",
-                "Global 2 m temp",
-                "Surface temperature worldwide",
-                false,
-            ),
-            (
-                FL::GlobalWind10m,
-                "Models",
-                "Global 10 m wind",
-                "Surface wind worldwide",
-                false,
-            ),
-            (
-                FL::GlobalPrecip,
-                "Models",
-                "Global moisture",
-                "Precipitable water (GFS) or total precipitation (ECMWF)",
-                false,
-            ),
-            (
-                FL::ModelDiff,
-                "Models",
-                "Model difference",
-                "Where two models disagree \u{2014} pick the field in layer options",
-                false,
-            ),
-            (
-                FL::Hrrr,
-                "Models",
-                "HRRR future radar",
-                "Forecast radar picture for the next 18 hours (not observed)",
-                true,
-            ),
-            (
-                FL::UpdraftHelicity,
-                "Models",
-                "Future rotation tracks",
-                "Where storms are forecast to rotate \u{2014} scrub the timeline to extend the swath",
-                true,
-            ),
-            (
-                FL::SnowAnalysis,
-                "National",
-                "Snowfall analysis",
-                "How much snow actually fell \u{2014} pick the window in layer options",
-                false,
-            ),
-            (
-                FL::Snowfall,
-                "Models",
-                "Forecast snowfall",
-                "How much snow is forecast to pile up \u{2014} scrub the timeline to add hours",
-                false,
-            ),
-            (
-                FL::Smoke,
-                "Models",
-                "Wildfire smoke",
-                "Forecast smoke near the ground, from active fires",
-                false,
-            ),
-            (
-                FL::Cape,
-                "Models",
-                "CAPE",
-                "How much fuel the atmosphere has for storms",
-                false,
-            ),
-            (
-                FL::Srh,
-                "Models",
-                "Storm-relative helicity",
-                "How much spin the wind profile can feed a storm",
-                false,
-            ),
-        ] {
-            let on = self.fields.get(&layer).is_some_and(|s| s.show);
-            push(
-                label,
-                category,
-                desc,
-                common,
-                PaletteAction::ToggleField(layer),
-                Some(on),
-            );
-        }
-        for k in ContourKind::ALL {
-            let label = format!("Contours: {}", k.label());
-            let on = self.contour_kind == k;
-            push(
-                &label,
-                "Models",
-                "Draw this forecast field as labeled contour lines",
-                false,
-                PaletteAction::SetContours(k),
-                Some(on),
-            );
-        }
-
-        // --- Severe / obs / reference toggles. ---
-        for (t, category, label, desc, common) in [
-            (
-                T::Cells,
-                "Severe",
-                "Storm cells",
-                "Mark each storm the radar is tracking",
-                true,
-            ),
-            (
-                T::Alerts,
-                "Severe",
-                "NWS alerts",
-                "Official warning and watch polygons",
-                true,
-            ),
-            (
-                T::Couplets,
-                "Severe",
-                "Rotation couplets",
-                "Flag tight rotation that could produce a tornado",
-                true,
-            ),
-            (
-                T::Tbss,
-                "Severe",
-                "Hail spikes (TBSS)",
-                "Flag three-body scatter spikes \u{2014} near-proof of large hail in the core \
-                 they point away from",
-                false,
-            ),
-            (
-                T::ZdrColumns,
-                "Severe",
-                "ZDR columns",
-                "Flag rain carried above the freezing level \u{2014} an updraft proxy that \
-                 deepens before a storm intensifies",
-                false,
-            ),
-            (
-                T::Tds,
-                "Severe",
-                "TDS detection",
-                "Flag lofted debris — a tornado is likely on the ground",
-                true,
-            ),
-            (
-                T::StormReports,
-                "Severe",
-                "Storm reports (LSR)",
-                "What people on the ground actually reported today",
-                true,
-            ),
-            (
-                T::AlertPanel,
-                "Severe",
-                "Active alerts list",
-                "Every alert in view, worst first (the sidebar's Alerts tab)",
-                true,
-            ),
-            (
-                T::Tracks,
-                "Severe",
-                "SCIT forecast tracks",
-                "Where each tracked storm is projected to go",
-                false,
-            ),
-            (
-                T::ArrivalCones,
-                "Severe",
-                "Arrival-time cones",
-                "When a storm is expected to reach points downstream",
-                false,
-            ),
-            (
-                T::Nowcast,
-                "Severe",
-                "Nowcast (echo extrapolation)",
-                "Short-range radar forecast by sliding echoes forward",
-                false,
-            ),
-            (
-                T::Mds,
-                "Severe",
-                "Mesoscale discussions",
-                "SPC's notes on where watches may be issued next",
-                false,
-            ),
-            (
-                T::ProbSevere,
-                "Severe",
-                "ProbSevere",
-                "Per-storm probability of severe weather, from NOAA/CIMSS",
-                false,
-            ),
-            (
-                T::Metar,
-                "Obs",
-                "Surface obs (METAR)",
-                "Temperature, dewpoint and wind at airports",
-                true,
-            ),
-            (
-                T::Webcams,
-                "Obs",
-                "Webcams (FAA + Windy)",
-                "Look at the sky through a real camera \u{2014} FAA airports, plus the Windy \
-                 network worldwide with a key in Settings",
-                false,
-            ),
-            (
-                T::Fires,
-                "Severe",
-                "Wildfires (WFIGS)",
-                "Active fire perimeters and incident points from the interagency fire feed",
-                false,
-            ),
-            (
-                T::Aqi,
-                "Obs",
-                "Air quality (AirNow)",
-                "EPA AQI at every monitor in view \u{2014} needs a free AirNow key in Settings",
-                false,
-            ),
-            (
-                T::Stations,
-                "Obs",
-                "Live station cards",
-                "Cameras and live telemetry from surface stations, one floating card each",
-                false,
-            ),
-            (
-                T::Dat,
-                "Obs",
-                "Damage surveys (NWS DAT)",
-                "What the survey crews found on the ground, rated point by point",
-                false,
-            ),
-            (
-                T::Spotters,
-                "Obs",
-                "Spotter Network",
-                "Live positions of storm spotters near the radar",
-                true,
-            ),
-            (
-                T::Gauges,
-                "Obs",
-                "River gauges (NWPS)",
-                "River levels and flood stage",
-                false,
-            ),
-            (
-                T::Sensors,
-                "Obs",
-                "Sensor dashboard",
-                "Current conditions and 24-hour trends at the nearest station",
-                false,
-            ),
-            (
-                T::Hodo,
-                "Obs",
-                "VAD hodograph",
-                "How the wind turns with height above the radar",
-                false,
-            ),
-            (
-                T::Tropical,
-                "Obs",
-                "Tropical (NHC)",
-                "Hurricane tracks and forecast cones",
-                false,
-            ),
-            (
-                T::Mping,
-                "Obs",
-                "Crowd reports (mPING)",
-                "What people outside say is falling: rain, snow, sleet, freezing rain",
-                false,
-            ),
-            (
-                T::Pireps,
-                "Obs",
-                "Pilot reports (PIREPs)",
-                "What pilots actually flew through: turbulence, icing, cloud tops",
-                false,
-            ),
-            (
-                T::Recon,
-                "Obs",
-                "Recon flight track",
-                "Hurricane-hunter observations: flight-level and surface wind, measured",
-                false,
-            ),
-            (
-                T::Aviation,
-                "Obs",
-                "Aviation (SIGMET/AIRMET)",
-                "Hazard areas for pilots: turbulence, icing, low ceilings",
-                false,
-            ),
-            (
-                T::Tfr,
-                "Reference",
-                "Flight restrictions (TFR)",
-                "Airspace you may not fly through: fires, stadiums, VIP movements, launches",
-                false,
-            ),
-            (
-                T::Blockage,
-                "Reference",
-                "Beam blockage (terrain)",
-                "Shade where terrain cuts into this tilt's beam \u{2014} the radar's blind spots",
-                false,
-            ),
-            (
-                T::RadarSites,
-                "Reference",
-                "Radar sites",
-                "Show every NEXRAD site; click one to switch radars",
-                true,
-            ),
-            (
-                T::Wind,
-                "Models",
-                "Wind (animated)",
-                "HRRR 10 m wind as drifting particles \u{2014} forecast output, CONUS only",
-                true,
-            ),
-            (
-                T::GlmLightning,
-                "Severe",
-                "Satellite lightning (GLM)",
-                "Total lightning (optical) — every flash the GOES mapper sees, in-cloud included, \
-                 fading as it ages. The CG density layer is the ground-strike half.",
-                true,
-            ),
-            (
-                T::Fronts,
-                "Reference",
-                "Surface fronts (H/L)",
-                "The cold, warm and stationary fronts from the national weather map",
-                true,
-            ),
-            (
-                T::RangeRings,
-                "Reference",
-                "Range rings",
-                "Distance rings around the radar, every 50 km",
-                true,
-            ),
-            (
-                T::LinkCameras,
-                "Reference",
-                "Link pane cameras",
-                "Pan and zoom every pane together",
-                false,
-            ),
-            (
-                T::MiniLoop,
-                "Reference",
-                "Mini loop window",
-                // Honest about the caveat rather than promising something the compositor will
-                // not do — see `mini_loop_viewport`.
-                if cfg!(unix) && std::env::var_os("WAYLAND_DISPLAY").is_some() {
-                    "A small window showing the active pane (Wayland cannot keep it on top)"
-                } else {
-                    "A small always-on-top window showing the active pane"
-                },
-                false,
-            ),
-        ] {
-            let on = *self.overlay_flag(t);
-            push(
-                label,
-                category,
-                desc,
-                common,
-                PaletteAction::ToggleOverlay(t),
-                Some(on),
-            );
-        }
-        push(
-            "Cycle basemap",
-            "Reference",
-            "Switch the map underneath the radar",
-            true,
-            PaletteAction::CycleBasemap,
-            None,
-        );
-        push(
-            "Open this view in Windy",
-            "Reference",
-            "Open windy.com in your browser, looking at the same place",
-            true,
-            PaletteAction::OpenInWindy,
-            None,
-        );
-        push(
-            "Copy link to this view",
-            "Reference",
-            "A link to this site, place, zoom, time and product \u{2014} opens Hook Echo here",
-            true,
-            PaletteAction::CopyViewLink,
-            None,
-        );
-        push(
-            "Save workspace",
-            "Reference",
-            "Remember this pane layout \u{2014} sites, products, tilts, overlays \u{2014} to restore later",
-            true,
-            PaletteAction::SaveWorkspace,
-            None,
-        );
-        for (i, ws) in self.settings.workspaces.iter().enumerate() {
-            push(
-                &format!("Workspace: {}", ws.name),
-                "Reference",
-                "Restore this saved pane layout",
-                true,
-                PaletteAction::ApplyWorkspace(i),
-                None,
-            );
-        }
-        push(
-            "Mute audio alerts",
-            "Alerts",
-            "Silence every chime and spoken warning without changing your sound choices",
-            true,
-            PaletteAction::ToggleMute,
-            Some(self.settings.mute_alerts),
-        );
-        push(
-            "Sidebar",
-            "Reference",
-            "This panel \u{2014} hide it and the map runs to the left edge",
-            false,
-            PaletteAction::ToggleSidebar,
-            Some(!self.settings.hide_sidebar),
-        );
-        push(
-            "Timeline bar",
-            "Reference",
-            "The transport strip under the map \u{2014} hide it for an edge-to-edge map",
-            false,
-            PaletteAction::ToggleToolbar,
-            Some(!self.settings.hide_toolbar),
-        );
-        push(
-            "About Hook Echo-WX",
-            "Reference",
-            "Version, links, and whether a newer release is out",
-            false,
-            PaletteAction::OpenWindow(AppWindow::About),
-            None,
-        );
-
-        // --- Tools, windows, panes. ---
-        let tool = self.tool;
-        for (t, label, desc, common) in [
-            (
-                MapTool::Interrogate,
-                "Tool: Interrogate",
-                "Click anywhere to read the exact radar value",
-                true,
-            ),
-            (
-                MapTool::Measure,
-                "Tool: Measure",
-                "Drag to measure distance and bearing",
-                true,
-            ),
-            (
-                MapTool::Marker,
-                "Tool: Drop marker",
-                "Save a place — home, work, where you're headed",
-                true,
-            ),
-            (
-                MapTool::Forecast,
-                "Tool: Point forecast",
-                "Tap anywhere for that spot's 7-day and hourly forecast",
-                true,
-            ),
-            (
-                MapTool::Sounding,
-                "Tool: Sounding",
-                "Click a point for the model profile plus the nearest balloon sounding",
-                false,
-            ),
-            (
-                MapTool::CrossSection,
-                "Tool: Cross-section",
-                "Drag a line to slice the storm vertically",
-                false,
-            ),
-            (
-                MapTool::Chase,
-                "Tool: Set chase location",
-                "Tell the app where you are, for the chase readout",
-                false,
-            ),
-            (
-                MapTool::Climatology,
-                "Tool: Tornado climatology",
-                "How often tornadoes have hit this spot historically",
-                false,
-            ),
-            (
-                MapTool::AlertZone,
-                "Tool: Watch zone",
-                "Draw an area that alerts when a warning polygon touches it",
-                false,
-            ),
-            (
-                MapTool::Draw,
-                "Tool: Draw",
-                "Scribble on the map — circle the storm you're talking about",
-                false,
-            ),
-        ] {
-            push(
-                label,
-                "Tools",
-                desc,
-                common,
-                PaletteAction::Tool(t),
-                Some(tool == t),
-            );
-        }
-        for (w, label, desc, common) in [
-            (
-                W::Site,
-                "Radar site…",
-                "Pick which radar you're watching",
-                true,
-            ),
-            (
-                W::Settings,
-                "Settings…",
-                "Theme, units, time display, alert sounds",
-                true,
-            ),
-            (
-                W::Markers,
-                "Location markers…",
-                "Manage your saved places and their alerts",
-                false,
-            ),
-            (
-                W::Events,
-                "Event library…",
-                "Jump to a famous storm and watch it replay",
-                false,
-            ),
-            (
-                W::Digest,
-                "Storm digest…",
-                "A plain-language summary of what's happening now",
-                false,
-            ),
-            (
-                W::Afd,
-                "Forecast discussion (AFD)…",
-                "What the local forecast office is writing",
-                false,
-            ),
-            (
-                W::Placefiles,
-                "Placefile manager…",
-                "Add GRLevelX placefile overlays",
-                false,
-            ),
-            (
-                W::LayerManager,
-                "Layer manager…",
-                "Reorder and set opacity for every layer",
-                false,
-            ),
-            (
-                W::Palettes,
-                "Color-table editor…",
-                "Change the colors a product is drawn with",
-                false,
-            ),
-            (
-                W::StormTable,
-                "Storm attributes…",
-                "Every tracked storm in one sortable table \u{2014} hail size, tops, rotation",
-                true,
-            ),
-            (
-                W::AlertRules,
-                "Alert rules\u{2026}",
-                "Tell the app what is worth interrupting you for",
-                false,
-            ),
-            (
-                W::Glossary,
-                "Glossary\u{2026}",
-                "What a TDS, a hail spike or a ZDR column actually is",
-                false,
-            ),
-            (
-                W::Verify,
-                "Warning verification…",
-                "Score an office's warnings against what actually happened",
-                false,
-            ),
-            (
-                W::Cappi,
-                "CAPPI slice…",
-                "See the storm at one constant altitude",
-                false,
-            ),
-            (
-                W::Volume3d,
-                "3D volume…",
-                "Rotate the storm in three dimensions",
-                false,
-            ),
-            (
-                W::Climatology,
-                "Tornado climatology…",
-                "Historical tornado tracks for this area",
-                false,
-            ),
-            (W::Wizard, "Setup wizard…", "Re-run first-time setup", false),
-            (
-                W::Tour,
-                "Take the tour…",
-                "A 60-second walk through the app's controls",
-                false,
-            ),
-        ] {
-            // The raymarch samples a `texture_3d<u32>`, which WebGL2 does not guarantee; on wasm
-            // the entry would open a black window.
-            // ponytail: revisit when the web build is webgpu-only.
-            if cfg!(target_arch = "wasm32") && w == W::Volume3d {
-                continue;
-            }
-            let on = None;
-            push(
-                label,
-                "Tools",
-                desc,
-                common,
-                PaletteAction::OpenWindow(w),
-                on,
-            );
-        }
-        push(
-            "Compare 4 tilts",
-            "Tools",
-            "Four panes of this product at four heights, cameras linked",
-            true,
-            PaletteAction::AllTilts,
-            None,
-        );
-        let panes = self.views.len();
-        for n in [1usize, 2, 4] {
-            push(
-                &format!("{n} pane{}", if n == 1 { "" } else { "s" }),
-                "Tools",
-                "Split the window to watch several radars or products at once",
-                false,
-                PaletteAction::SetPanes(n),
-                Some(panes == n),
-            );
-        }
-        push(
-            "Reload",
-            "Tools",
-            "Fetch the latest data again",
-            true,
-            PaletteAction::Reload,
-            None,
-        );
-        push(
-            "Jump to live",
-            "Tools",
-            "Snap back to the newest scan",
-            true,
-            PaletteAction::GoLive,
-            None,
-        );
-        push(
-            "Instant replay (DVR)",
-            "Tools",
-            "Replay the scans already in memory",
-            false,
-            PaletteAction::InstantReplay,
-            None,
-        );
-        out
     }
 
     /// Run one registry action. Every surface (drawer, pills, mobile sheets) routes through it.
@@ -11101,52 +9420,6 @@ impl HookEchoApp {
                 self.views[idx].error = Some(e.to_string());
                 (None, false)
             }
-        }
-    }
-
-    /// Offer back the report from a run that panicked. Once, with a Copy button — the point is
-    /// that a crash the user can only describe as "it closed" becomes something they can paste
-    /// into an issue.
-    fn crash_report_window(&mut self, ctx: &egui::Context) {
-        let Some(report) = self.crash_report.clone() else {
-            return;
-        };
-        let mut open = true;
-        let mut dismiss = false;
-        egui::Window::new("Hook Echo-WX closed unexpectedly last time")
-            .open(&mut open)
-            .collapsible(false)
-            .default_size([560.0, 320.0])
-            .show(ctx, |ui| {
-                ui.label("Here is what it left behind. Nothing in it identifies you.");
-                ui.add_space(4.0);
-                egui::ScrollArea::vertical()
-                    .max_height(220.0)
-                    .show(ui, |ui| {
-                        ui.add(
-                            egui::TextEdit::multiline(&mut report.as_str())
-                                .font(egui::TextStyle::Monospace)
-                                .desired_width(f32::INFINITY),
-                        );
-                    });
-                ui.add_space(4.0);
-                ui.horizontal(|ui| {
-                    if ui.button("Copy report").clicked() {
-                        ui.ctx().copy_text(report.clone());
-                    }
-                    if ui.button("Dismiss").clicked() {
-                        dismiss = true;
-                    }
-                    ui.hyperlink_to(
-                        "Open an issue",
-                        "https://github.com/d4vid87/hookecho/issues/new",
-                    );
-                });
-            });
-        if dismiss || !open {
-            #[cfg(not(target_arch = "wasm32"))]
-            crate::crash::dismiss();
-            self.crash_report = None;
         }
     }
 
@@ -15065,198 +13338,6 @@ impl HookEchoApp {
         }
     }
 
-    /// Bottom-right info chip: zoom, cursor position, DVR depth, and the active tool's hint.
-    ///
-    /// This is what the docked status bar used to hold. A full-width bar for four short readouts
-    /// cost a strip of map on every frame; the chip floats over the map instead, and the site and
-    /// volume time it also carried now live in the timeline pill where the clock belongs.
-    /// The way back to a hidden sidebar: one floating button in the corner it used to occupy.
-    /// Nothing else on the map reaches the layer list, so this is not optional chrome.
-    fn sidebar_button(&mut self, ctx: &egui::Context) {
-        if !self.settings.hide_sidebar {
-            return;
-        }
-        egui::Area::new(egui::Id::new("sidebar_button"))
-            .constrain_to(self.chrome_rect)
-            .anchor(egui::Align2::LEFT_TOP, egui::vec2(10.0, 10.0))
-            .show(ctx, |ui| {
-                crate::ui::style::glass(ui, 238).show(ui, |ui| {
-                    if ui
-                        .add(
-                            egui::Button::new(
-                                egui::RichText::new(egui_phosphor::regular::LIST).size(18.0),
-                            )
-                            .min_size(egui::vec2(30.0, 30.0))
-                            .fill(egui::Color32::TRANSPARENT)
-                            .stroke(egui::Stroke::NONE),
-                        )
-                        .on_hover_text("Show the sidebar")
-                        .clicked()
-                    {
-                        self.settings.hide_sidebar = false;
-                        self.settings.save();
-                    }
-                });
-            });
-    }
-
-    fn info_chip(&mut self, ctx: &egui::Context) {
-        use crate::ui::style;
-        use egui_phosphor::regular as ph;
-        // Icon, tool name, and what to do with it — an armed tool should look armed, not leave a
-        // sentence of instructions as the only sign anything changed.
-        let (glyph, name, hint) = match self.tool {
-            MapTool::Measure => (ph::RULER, "Measure", "click two points"),
-            MapTool::Marker => (ph::MAP_PIN, "Drop marker", "click the map"),
-            MapTool::CrossSection => (ph::CHART_LINE, "Cross-section", "click two points"),
-            MapTool::Sounding => (ph::THERMOMETER_SIMPLE, "Sounding", "click a point"),
-            MapTool::Forecast => (ph::CLOUD_SUN, "Forecast", "click a point"),
-            MapTool::Chase => (ph::CROSSHAIR, "Chase", "click your location"),
-            MapTool::Climatology => (ph::TORNADO, "Climatology", "click a point"),
-            MapTool::Draw => (ph::PENCIL_SIMPLE, "Draw", "drag to scribble"),
-            MapTool::AlertZone => (
-                ph::POLYGON,
-                "Watch zone",
-                "click each corner · double-click to close",
-            ),
-            MapTool::Interrogate => ("", "", ""),
-        };
-        // Nothing armed, nothing to say. The chip used to be a permanent readout of the cursor's
-        // lat/lon and the zoom level — numbers nobody was reading, in the corner where the map is.
-        if hint.is_empty() {
-            return;
-        }
-        let accent = crate::theme::accent(self.settings.theme);
-        // An armed tool changes what a click means, so the cursor says so over the whole map.
-        ctx.set_cursor_icon(egui::CursorIcon::Crosshair);
-        // Escape disarms, the same way it closes every other transient thing.
-        if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
-            self.tool = MapTool::Interrogate;
-            return;
-        }
-        egui::Area::new(egui::Id::new("info_chip"))
-            .constrain_to(self.chrome_rect)
-            .anchor(
-                egui::Align2::RIGHT_BOTTOM,
-                egui::vec2(-14.0, style::LANE_BOTTOM_CHIP),
-            )
-            // The draw tool's chip carries buttons; the rest are read-only hints that must never
-            // swallow a click meant for the map underneath.
-            .interactable(self.tool == MapTool::Draw)
-            .show(ctx, |ui| {
-                style::glass(ui, 238).show(ui, |ui| {
-                    ui.horizontal(|ui| {
-                        ui.label(egui::RichText::new(glyph).size(15.0).color(accent));
-                        ui.label(
-                            egui::RichText::new(name)
-                                .size(style::FONT_BASE)
-                                .color(accent)
-                                .strong(),
-                        );
-                        ui.label(egui::RichText::new(hint).size(style::FONT_SM).weak());
-                        ui.label(
-                            egui::RichText::new("· Esc cancels")
-                                .size(style::FONT_SM)
-                                .weak(),
-                        );
-                        if self.tool != MapTool::Draw {
-                            return;
-                        }
-                        ui.separator();
-                        for c in DRAW_COLORS {
-                            let sel = self.draw_color == c;
-                            let (rect, resp) = ui
-                                .allocate_exact_size(egui::vec2(18.0, 18.0), egui::Sense::click());
-                            let p = ui.painter_at(rect);
-                            p.circle_filled(rect.center(), 7.0, c);
-                            if sel {
-                                p.circle_stroke(
-                                    rect.center(),
-                                    8.5,
-                                    egui::Stroke::new(1.5, egui::Color32::WHITE),
-                                );
-                            }
-                            if resp.clicked() {
-                                self.draw_color = c;
-                            }
-                        }
-                        ui.separator();
-                        if ui
-                            .add_enabled(!self.strokes.is_empty(), egui::Button::new("Undo"))
-                            .clicked()
-                        {
-                            self.strokes.pop();
-                        }
-                        if ui
-                            .add_enabled(!self.strokes.is_empty(), egui::Button::new("Clear"))
-                            .clicked()
-                        {
-                            self.strokes.clear();
-                        }
-                    });
-                });
-            });
-    }
-
-    /// Bottom-center error chip: the active pane's fetch error, auto-hiding after ~6 seconds.
-    ///
-    /// ponytail: one chip, newest error wins, no toast queue — add a queue if overlapping errors
-    /// from different panes turn out to matter.
-    fn error_chip(&mut self, ctx: &egui::Context) {
-        const HOLD_SECS: f64 = 6.0;
-        let now = ctx.input(|i| i.time);
-        if let Some(e) = self.views[self.active].error.clone() {
-            if self.error_chip.as_ref().is_none_or(|(prev, _)| *prev != e) {
-                self.error_chip = Some((e, now));
-            }
-        }
-        let Some((msg, since)) = self.error_chip.clone() else {
-            return;
-        };
-        if now - since > HOLD_SECS {
-            self.error_chip = None;
-            return;
-        }
-        ctx.request_repaint_after(std::time::Duration::from_millis(500));
-        let chip = egui::Area::new(egui::Id::new("error_chip"))
-            .constrain_to(self.chrome_rect)
-            .anchor(
-                egui::Align2::CENTER_BOTTOM,
-                egui::vec2(0.0, crate::ui::style::LANE_BOTTOM_CHASE),
-            )
-            // Interactable so the message can be read and copied: six seconds is not long enough
-            // to transcribe a URL out of a failure. The chip is small and sits in the bottom lane,
-            // so what it costs the map underneath is that strip.
-            .interactable(true)
-            .show(ctx, |ui| {
-                crate::ui::style::glass(ui, 246)
-                    .stroke(egui::Stroke::new(
-                        1.0,
-                        egui::Color32::from_rgb(230, 100, 100).gamma_multiply(0.8),
-                    ))
-                    .show(ui, |ui| {
-                        ui.add(
-                            egui::Label::new(
-                                egui::RichText::new(&msg)
-                                    .size(crate::ui::style::FONT_BASE)
-                                    .color(egui::Color32::from_rgb(240, 150, 150)),
-                            )
-                            .sense(egui::Sense::click()),
-                        )
-                        .on_hover_text("Click to copy")
-                    })
-            });
-        let resp = &chip.inner.inner;
-        if resp.hovered() {
-            // Hold while it is being read.
-            self.error_chip = Some((msg.clone(), now));
-        }
-        if resp.clicked() {
-            ctx.copy_text(msg);
-            self.toast(ToastKind::Success, "Error copied");
-            self.error_chip = None;
-        }
-    }
 }
 
 /// Convert a binned sweep into a GPU upload with its world-space bounding box.
