+++ crates/hookecho/src/app/chrome/mod.rs
@@ -1,0 +1,10 @@
+//! Desktop chrome: everything drawn around the map. Split out of app.rs so each surface is its
+//! own file; the bodies moved verbatim.
+
+mod chips;
+mod registry;
+mod sidebar;
+mod timeline;
+mod windows;
+
+use super::*;
